Fixed couple of Econolite decoder issues
This change fixes two minor but important issues in the decoder.
First every second event of each file was beeing lost. This was because "if (!firstEventEntered)" statement had an else statement which was actually always true during the second event. This would insert first event and drop the second.
Second important issue resolved in this changeset is that the format of the Econolite DAT file is now (eventCode,EventParam,TimeofEvent). The code reading actually assumes the slightly different format (TimeOfEvent, EventCode,EventParam). This would cause a slight difference in timestamps that is usally hard to notice but is present.