added ability to handle compressed files (*.datZ format in Econolite)
Added the capability to decompress the files that are compressed and then pass the decompressed stream into decoder for further processing. The decompression code should live in another file.
Please be aware that we do not use this code base and have not tested this changes in this exact code base but in a similar decoder. Test with good known files before using.