Fix for Data Aggregation Memory Leak
Data aggregation had a memory leak due to Entity Framework and the Factory pattern. The ability to supply a context and turning off change tracking fixed the issue. Also a bug in signal phase was discovered where if the user did not put in mph on a detector it would fail.